Fix incorrect var_end_comment() reporting
var_end_comment() provides a human-readable comment appropriate for
annotating an engine line added as a variation. Typically this will be
in the format "eval/depth" where eval is the engine evaluation and depth
is the search depth reached to produce the variation. However, if the
engine variation ends in checkmate, "eval" would be set to None. This
was causing the function to return string values such as "None/17" if
the variation ends in checkmate.

Additionally, var_end_comment() would return a search depth for a
variation that ends the game, which is usually irrelevant.

This commit reworks var_end_comment() to return "" if the variation ends
in checkmate, and not to report search depth.