Check engine variant support before analyzing
Attempting to analyze a variant game with an engine that does not
support the variant tends to produce ugly results. This code attempts to
poll the engine to check if the UCI_Variant option is provided, and if
so, if the current game's variant is in that list. An exception is
raised if either check fails.