Remove superfluous 'invert' parameter
This parameter was being used to invert the result when the evaluation
context shifts from one player to the other. But since the returned
value of eval_absolute() is inverted or not accordnig to board.turn, the
same effect can be achieved by simply passing 'not board.turn' to
eval_human() when we want an inverted return value.