Division by zero error if ply_count is zero
When running the program, I sometimes ran into two different division by zero problems, caused by the fact that ply_count = 0. My solution is that the minimum of the division with ply_count is 1, so a division by zero is not possible.  I suppose it would be better to ensure that ply_count never is zero, but I'm not sure how hard that is as I haven't really looked into the problem.